pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.49.0",
    "email-validator>=2.2.0",
    "flask-cors>=5.0.1",
    "flask>=3.1.0",
    "flask-login>=0.6.3",
    "flask-restful>=0.3.10",
    "flask-sqlalchemy>=3.1.1",
    "flask-wtf>=1.2.2",
    "gunicorn>=23.0.0",
    "mistralai>=1.6.0",
    "openai>=1.72.0",
    "psycopg2-binary>=2.9.10",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.0",
    "schedule>=1.2.2",
    "sqlalchemy>=2.0.40",
    "werkzeug>=3.1.3",
    "requests>=2.32.3",
    "psutil>=7.0.0",
]
